Promote methods from int to ptrdiff_t to allow extension to 64-bits.

+++ cocoa/ScintillaCocoa.h
@@ -107,8 +107,8 @@ private:
 
 	bool GetPasteboardData(NSPasteboard *board, SelectionText *selectedText);
 	void SetPasteboardData(NSPasteboard *board, const SelectionText &selectedText);
-	int TargetAsUTF8(char *text);
-	int EncodedFromUTF8(char *utf8, char *encoded) const;
+	ptrdiff_t TargetAsUTF8(char *text);
+	ptrdiff_t EncodedFromUTF8(char *utf8, char *encoded) const;
 
 	int scrollSpeed;
 	int scrollTicks;
